4.15-4.21

+++ part4/controllers/blogs.js
@@ -13,7 +13,6 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  console.log(decodedToken)
   if (!decodedToken) {
     return response.status(401).json({ error: 'token invalid' })
   }
@@ -34,8 +33,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const userid = decodedToken.id
+  const blog = await Blog.findById(request.params.id)
+
+  if ( !blog ) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  console.log('blog retrieved',blog)
+  console.log('userid', userid )
+  if ( blog.user.toString() === userid.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    return response.status(401).json({ error: 'unauthorized' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
